requirements.txt
markitdown
lingua-language-detector
spacy
